requirements.txt
requests
python-dotenv
openai
anthropic
tenacity